requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
django==5.2.8
django-cors-headers==4.7.0
django-filter==25.2
django-mssql-backend==2.8.1
django-nested-admin==4.1.6
djangorestframework==3.15.2
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
et-xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.27.2
idna==3.10
inflection==0.5.1
jaydebeapi==1.2.3
jpype1==1.5.2
kombu==5.5.4
# mssql-django==1.5
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pyjwt==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.1.0
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
